requirements.txt
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
asyncmy==0.2.8
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.110.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.3.0
pydantic_core==2.6.3
python-jose==3.3.0
pytz==2023.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.23.2
Werkzeug==2.3.7
